+++ packages/lumx-react/src/components/link/Link.stories.tsx
@@ -28,6 +28,7 @@ export const AllStates = { ...stories.AllStates };
 export const AllTypography = { ...stories.AllTypography };
 export const AllColors = { ...stories.AllColors };
 export const ParentTypographyAndColor = { ...stories.ParentTypographyAndColor };
+export const InParagraph = { ...stories.InParagraph };
 
 /** React-only: Check wrapping a block with a Link */
 export const WrappingBlock = {